test(solver): shard relation policy cache tests
